Revert "Task bootstrap and apply Auth0"
