+++ .eslintrc.json
@@ -5,11 +5,9 @@
     },
     "extends": [
         "eslint:recommended",
-        "plugin:jsx-a11y/recommended",
         "plugin:react/recommended",
         "airbnb",
-        "airbnb/hooks",
-        "prettier"
+        "airbnb/hooks"
     ],
     "parserOptions": {
         "ecmaFeatures": {
@@ -19,15 +17,9 @@
         "sourceType": "module"
     },
     "plugins": [
-        "react",
-        "jsx-a11y",
-        "prettier"
+        "react"
     ],
     "rules": {
-        "react/function-component-definition": "off",
-        "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }],
-        "react/forbid-prop-types": 0,
-        "react/require-default-props": "off",
-        "react/prop-types": "off"
+        "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }]
     }
 }
